moved filter toolbar to shared folder
