+++ clients/ui/frontend/src/app/pages/modelRegistry/screens/RegisteredModelsArchive/RegisteredModelsArchiveListView.tsx
@@ -10,7 +10,7 @@ import {
   initialModelRegistryFilterData,
   modelRegistryFilterOptions,
 } from '~/app/pages/modelRegistry/screens/const';
-import FilterToolbar from '~/app/pages/modelRegistry/screens/components/FilterToolbar';
+import FilterToolbar from '~/app/shared/components/FilterToolbar';
 import ThemeAwareSearchInput from '~/app/pages/modelRegistry/screens/components/ThemeAwareSearchInput';
 import RegisteredModelsArchiveTable from './RegisteredModelsArchiveTable';
